[UBDEV-771] Fix handling of UUIDs.
UUID from H2O's frame were not transfered to Spark RDD schema in
proper form.

Now, low,high bits are read from H2O chunk, UUID is created, and saved
into RDD.

Warning: It is not fully efficient implementation since it recreates UUID object
for each row, but UUIDs are not used so often. So optimize it later if
necessary.