[PUBDEV-928] Explicitly generate 0-length chunks.
Original code expected that SQL query causes generation of empty chunks, which
was not true on all test machines since number of H2O chunks varies based on number of available CPUs.

The fix forces number of chunks to be (nrows+1) which means that there will be at least one 0-length chunk.