Added Dataset to H2O spart interface. Products (aka case classes) can contain multiple primitive or optional components. Optional are of type Option[T], where T is primitive.
Serialization writes options as undefined.
Deserialization transforms undefined to None and legit values x to Some(x) - when the field is optional. When the field is primitive, undefined values mean the whole row is undefined, and it is being skipped. Any other transformation errors also lead to skipping offending rows.