H2ODataset - added tests for trying reading into multiple case classes with the same dataset, and what happens. A bunch of changes in H2ORDD, to avoid meaningless exceptions; dealing with exceptions (just returning None).
Not all the tests pass yet; and that's as intended. As soon as I do something with the code, they'll pass.

Work in progress; another issue is the multiplicity of constructors. It does not make sense; given a Frame or Dataset (whatever), the builder can be provided once for the whole set. No need to do reflection per each row.