[SW-178][Prep for external backend] Fix the problem of getting wrong H2OConf in H2ORDD and H2ODataFrame.
Right now wrong configuration is returned in those 2 objects. Companion object creates new H2OConf but it doesn't reflect the state of the
the configuration on H2OContext. We need to use configuration provided by H2OContext

This is also reason why apply method on companion object for H2OConf is removed since it is returning wrong H2OConf