don't fail immediately if SPARK_HOME isn't set
Sometimes the Spark is installed via package manager, or some other
means, like, Cloudera parcells.  In this case the SPARK_HOME isn't set
explicitly, but `spark-shell` & `spark-submit` are available in the
`PATH`.  This patch adds for this case, and fail only

It also fixes explicit use of `FAT_JAR` variable instead of using
`FAT_JAR_FILE` that points to the full path to `FAT_JAR`.