[SW-289] Fix PySparkling numeric handling during conversions
This behaviour was there also on original backend
Even though we are convertingRDD containing doubles, some of the elements are actually integers. We need to convert all types to doubles in order to not break the backend since we specify the expected type and that type has to be the same for all elements in the RDD
Also distinguish between long and integer type.