[SW-376] Fix DataType and DecimalType conversions on external backend. (#235)
The DateType is internally handled as Long and the corresponsing internal java class for external backend  has to be marked as long as well
The DecimalType is special type which is not listed in supported types. The corresponding internal type needs to be double.
(cherry picked from commit 90adde0737f50bbd4296c5ff520e78e43bf0fbd9)