[SW-389] Specify timeout for reading/writing.
When timeout is reached, the exception is thrown. It is not catched on purpose since the spark will stop the job on the current executor and restart it. That should fix several intermittent issues together with  the problem of a few jobs never finishing ( the problem appeared and the confirmation never appeared)

(cherry picked from commit 5fe49694ac4f3c30ff8e38be0f5163a873b8636e)