[SW-504] Provides uber Sparkling Water Spark package (#352)
* [SW-504] Provides uber Sparkling Water Spark package

The motivation for this is to improve user experience. Right now, SW provides 4 modules with complex dependencies which need to be often hot-fixed (like  SW-475).
The goal of this improvement is to provide a single uber Sparkling Water package which can be used as input for Spark submit option `--packages`.

This package is being published into maven central and can be used with Spark submit:
```
$SPARK_HOME/bin/spark-submit --packages ai.h2o:sparkling-water-package_2.11:2.1.12 --class org.apache.spark.examples.h2o.CraigslistJobTitlesStreamingApp /dev/null
```

Note: right now it also provides a fixed for [SW-475]
(cherry picked from commit 435d57c34e07654fe0bd195254499990f4bebe97)