[SW-851] Properly type and name output columns of mojo pipeline
The columns are now properly typed as double. Also, we currently store
the prediction output in an array for each row, loosing the information
about the name of the output column.

This has been fixed as well, however the behaviour is disabled by
default. It will be enabled by default in Spark 2.4